fix(keypool): float64 rate limit, panic recovery, forbid threshold
- rate_limit: parse pointsSpentThisHour as float64 (FFLogs returns decimals)
- fflogs: guard MapToMemo against empty Fights/Composition; worker-level
  recover so one bad member doesn't kill the goroutine
- keypool: stagger probes 300ms to avoid OAuth burst 429;
  default Limit=5000 when FFLogs returns 0 for fresh OAuth clients;
  preserve Limit/ResetAt during Reconcile when FFLogs returns 0
- keypool: isInvalidClient detection + forbidInDB writes err_count=100
  synchronously, Load filters err_count<100 so permanently-broken donated
  keys stop costing a probe on every restart
- memo: POST /fight/ trailing slash; remove per-client rate.Limiter now
  that memo-server bypasses limits for priority>=100 clients
- remove .github/workflows/fly-deploy.yaml (manual deploy during testing)

+++ service/fflogs/rate_limit.go
@@ -12,10 +12,14 @@ query {
 }
 `
 
+// RateLimitData mirrors FFLogs' rateLimitData payload. All three fields
+// come back as JSON numbers with fractional parts (e.g. pointsSpentThisHour
+// can be `26.05`), so they must be float64 on the Go side. Integer casts
+// happen at the KeyState boundary in the pool.
 type RateLimitData struct {
-	LimitPerHour        int `json:"limitPerHour"`
-	PointsSpentThisHour int `json:"pointsSpentThisHour"`
-	PointsResetIn       int `json:"pointsResetIn"`
+	LimitPerHour        float64 `json:"limitPerHour"`
+	PointsSpentThisHour float64 `json:"pointsSpentThisHour"`
+	PointsResetIn       float64 `json:"pointsResetIn"`
 }
 
 type rateLimitResponse struct {
